fix(jobboard): Fix TypeError in workflow logging and dean stats query
- Cast changedby to int in change_status() to handle string user IDs
- Simplify dean approved count query to only check 'dean_approved' status
- Fixes TypeError: Argument #4 ($changedby) must be of type ?int, string given

+++ local/jobboard/classes/application.php
@@ -723,7 +723,7 @@ class application {
     public function change_status(string $newstatus, string $comments = '', ?int $changedby = null): void {
         global $DB, $USER;
 
-        $changedby = $changedby ?? $USER->id;
+        $changedby = (int) ($changedby ?? $USER->id);
 
         // Validate transition.
         if (!$this->can_transition_to($newstatus)) {
